Add `/api/instances/list` (#2199)
Add an API method for listing instances with
filtering and pagination. This method will be used
in the new Instances page in the UI. It is similar
to the deprecated `/api/pools/list_instances`
method, except it allows filtering by fleet and
not by pool.

Also add the `fleet_id` and `fleet_name` fields to
the instance model so that the Instances page in
the UI can display fleet names and provide links
to fleet details.